refactor(core): deduplicate whitespace count_tokens into shared core utility
The approximate whitespace token-count convention (max(1, len(text.split())))
existed twice with identical bodies: retrieval_qa._utils.count_tokens and
ingestion._splitting._count_tokens. Two copies of one convention invite
silent divergence.

Move the canonical helper into core.utils (the only package both retrieval_qa
and ingestion may depend on per ADR-0011), delete the ingestion private copy,
and point the retrieval_qa chunkers at the shared helper directly.

Closes #175
