+++ core/src/core/__init__.py
@@ -1,3 +1,3 @@
 """Shared core package for the Learning Hub monorepo."""
 
-__all__ = ["clients", "config", "database", "exceptions", "types"]
+__all__ = ["clients", "config", "database", "exceptions", "types", "utils"]
